+++ kubernetes-client/src/main/java/io/fabric8/kubernetes/client/dsl/Editable.java
@@ -18,10 +18,15 @@ package io.fabric8.kubernetes.client.dsl;
 import java.util.function.Consumer;
 import java.util.function.UnaryOperator;
 
+import io.fabric8.kubernetes.api.builder.Visitor;
+
 public interface Editable<T> {
 
     T edit(UnaryOperator<T> function);
 
+    T edit(Visitor... visitors);
+
+    <V> T edit(Class<V> visitorType, Visitor<V> visitor);
 
     T accept(Consumer<T> function);
 }