+++ kubernetes-client/src/main/java/io/fabric8/kubernetes/client/dsl/internal/batch/v1/JobOperationsImpl.java
@@ -15,6 +15,7 @@
  */
 package io.fabric8.kubernetes.client.dsl.internal.batch.v1;
 
+import io.fabric8.kubernetes.api.builder.Visitor;;
 import io.fabric8.kubernetes.api.model.Pod;
 import io.fabric8.kubernetes.api.model.autoscaling.v1.Scale;
 import io.fabric8.kubernetes.client.dsl.LogWatch;
@@ -24,6 +25,7 @@ import io.fabric8.kubernetes.client.dsl.base.OperationContext;
 import io.fabric8.kubernetes.client.utils.PodOperationUtil;
 import okhttp3.OkHttpClient;
 import io.fabric8.kubernetes.api.model.batch.Job;
+import io.fabric8.kubernetes.api.model.batch.JobBuilder;
 import io.fabric8.kubernetes.api.model.batch.JobList;
 import io.fabric8.kubernetes.client.Config;
 import io.fabric8.kubernetes.client.KubernetesClientException;
@@ -234,6 +236,12 @@ public class JobOperationsImpl extends HasMetadataOperation<Job, JobList, Scalab
     }
 
     return super.replace(job);
+  }
+
+
+  @Override
+  public Job edit(Visitor... visitors) {
+    return patch(new JobBuilder(getMandatory()).accept(visitors).build());
   }
 
   static Map<String, String> getJobPodLabels(Job job) {