+++ CHANGELOG.md
@@ -7,6 +7,7 @@
 #### Improvements
 
 #### Dependency Upgrade
+* edit methods, now accept visitors as vararg or lambda.
 
 #### New Features